+++ src/publisher/pubMqtt.h
@@ -140,8 +140,8 @@ class PubMqtt {
             if(!mClient.connected())
                 return;
 
-            char topic[MQTT_TOPIC_LEN + 2];
-            snprintf(topic, (MQTT_TOPIC_LEN + 2), "%s/%s", mCfgMqtt->topic, subTopic);
+            char topic[(MQTT_TOPIC_LEN << 1) + 2];
+            snprintf(topic, ((MQTT_TOPIC_LEN << 1) + 2), "%s/%s", mCfgMqtt->topic, subTopic);
             if(addTopic)
                 mClient.publish(topic, QOS_0, retained, payload);
             else