+++ src/utils/helper.cpp
@@ -9,7 +9,7 @@ namespace ah {
     void ip2Arr(uint8_t ip[], const char *ipStr) {
         memset(ip, 0, 4);
         char *tmp = new char[strlen(ipStr)+1];
-        strncpy(tmp, ipStr, strlen(ipStr)+1);
+        snprintf(tmp, strlen(ipStr), ipStr);
         char *p = strtok(tmp, ".");
         uint8_t i = 0;
         while(NULL != p) {